Coding style changes
Rename functions to conform to libaom's naming convention.

Change the keyword 'inline' to the macro 'INLINE'.

Declare the function filter_intra_predictor_sse4_1() as inline.

Remove an unneeded int16_t cast in xx_roundn_epi16_unsigned().

Tested:
test_libaom --gtest_filter=*FilterIntraPred* \
    --gtest_also_run_disabled_tests

BUG=aomedia:2880

+++ aom_dsp/x86/synonyms.h
@@ -101,7 +101,7 @@ static INLINE __m128i xx_roundn_epu32(__m128i v_val_d, int bits) {
 }
 
 static INLINE __m128i xx_roundn_epi16_unsigned(__m128i v_val_d, int bits) {
-  const __m128i v_bias_d = _mm_set1_epi16((int16_t)((1 << bits) >> 1));
+  const __m128i v_bias_d = _mm_set1_epi16((1 << bits) >> 1);
   const __m128i v_tmp_d = _mm_add_epi16(v_val_d, v_bias_d);
   return _mm_srai_epi16(v_tmp_d, bits);
 }
